refactor: refactor the main big api file into many compoenets inside view
