+++ backend/api/views/__init__.py
@@ -1,0 +1,23 @@
+from . import auth
+from . import users
+from . import events
+from . import tickets
+from . import dashboard
+from . import approval
+from . import feedback
+from . import notifications
+from . import public_profile
+from . import verification
+
+__all__ = [
+    "auth",
+    "users",
+    "events",
+    "tickets",
+    "dashboard",
+    "approval",
+    "feedback",
+    "notifications",
+    "public_profile",
+    "verification",
+]
